+++ scripts/build.mjs
@@ -68,6 +68,10 @@ const ENTRIES = [
 		entry: 'src/event-aggregator/settings.js',
 		outDir: path.resolve( ROOT, 'build/event-aggregator-settings' ),
 	},
+	{
+		entry: 'src/current-request/index.js',
+		outDir: path.resolve( ROOT, 'build/current-request' ),
+	},
 ];
 
 buildDashboards( {